Provide the Accept header in all requests
A recent update of the content negotiation in the GleSYS API introduced a non-backward compatible change where the response format could not be determined. This occurs when the `Accept` header as well as the `format` parameter is missing from the request and `Content-Type` is set to `application/x-www-form-urlencoded`.

To ensure that XML is returned by the GleSYS API the `Accept` header should be provided with the value of `application/xml`.